refactor: extract top model card
Made-with: Cursor
